chore(folder-structure): Move matomo to backend app
